Tweak string deserializer
PyUnicode_New needs to know the maximum code point to be placed in the
new unicode object. An upper bound for the maximum code point in a
utf-8 string is

1114111 if the highest byte in the string is >= 0xf0
65535 if the highest byte in the string is in [0xc4, 0xf0)
255 if the highest byte in the string is < 0xc4

Streamline the code by finding the highest byte in the string and
comparing it against 0xf0 and 0xc4. The rust compiler vectorizes
bytes().max() on x86_64 (PMAXUB) and arm64 (UMAX, UMAXV), so there is
no need to manually vectorize the operation using std::arch or
std::simd.

Signed-off-by: Emanuele Giaquinta <emanuele.giaquinta@gmail.com>